Parser files
This was originally done as a school assignment and contained a skeleton code for the header files which we had to use in order to make things work. The header file made defining of parsing nodes a lot harder for me but once I got it working, it felt glorious.